Chapter 6
Got the ratings to display on the screen. The text from the book so far initialised the shadow dom with an empty state, I have passed in the state from the data.js json file to ensure all the sections display correctly. I must be very careful with bracket types in JSX, it is easy to use the wrong one!